move NinSnesTrack into its own file
