+++ src/main/CMakeLists.txt
@@ -225,6 +225,7 @@ target_sources(vgmtranscore
     formats/NinSnes/NinSnesScannerPatterns.cpp
     formats/NinSnes/NinSnesSeq.cpp
     formats/NinSnes/NinSnesSeqIntelli.cpp
+    formats/NinSnes/NinSnesTrack.cpp
     formats/NinSnes/NinSnesTrackModulation.cpp
     formats/Org/OrgScanner.cpp
     formats/Org/OrgSeq.cpp
